fix(issues): fix issue de-duplication
* fix issue de-duplication
* refactor utilities
